+++ spec/dummy_app/app/models/no_object.rb
@@ -3,7 +3,7 @@
 # Demonstrates a table that omits the `object` column.
 class NoObject < ActiveRecord::Base
   has_paper_trail(
-    class_name: "NoObjectVersion",
+    versions: { class_name: "NoObjectVersion" },
     meta: { metadatum: 42 }
   )
   validates :letter, length: { is: 1 }, presence: true