+++ spec/models/custom_primary_key_record_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe CustomPrimaryKeyRecord, type: :model do
   describe "#versions" do
     it "returns instances of CustomPrimaryKeyRecordVersion", versioning: true do
       custom_primary_key_record = described_class.create!
-      custom_primary_key_record.update_attributes!(name: "bob")
+      custom_primary_key_record.update!(name: "bob")
       version = custom_primary_key_record.versions.last
       expect(version).to be_a(CustomPrimaryKeyRecordVersion)
       version_from_db = CustomPrimaryKeyRecordVersion.last