+++ spec/models/skipper_spec.rb
@@ -5,7 +5,7 @@ require "spec_helper"
 RSpec.describe Skipper, type: :model, versioning: true do
   it { is_expected.to be_versioned }
 
-  describe "#update_attributes!", versioning: true do
+  describe "#update!", versioning: true do
     context "updating a skipped attribute" do
       let(:t1) { Time.zone.local(2015, 7, 15, 20, 34, 0) }
       let(:t2) { Time.zone.local(2015, 7, 15, 20, 34, 30) }
@@ -13,7 +13,7 @@ RSpec.describe Skipper, type: :model, versioning: true do
       it "does not create a version" do
         skipper = Skipper.create!(another_timestamp: t1)
         expect {
-          skipper.update_attributes!(another_timestamp: t2)
+          skipper.update!(another_timestamp: t2)
         }.not_to(change { skipper.versions.length })
       end
     end
@@ -26,7 +26,7 @@ RSpec.describe Skipper, type: :model, versioning: true do
     context "without preserve (default)" do
       it "has no timestamp" do
         skipper = Skipper.create!(another_timestamp: t1)
-        skipper.update_attributes!(another_timestamp: t2, name: "Foobar")
+        skipper.update!(another_timestamp: t2, name: "Foobar")
         skipper = skipper.versions.last.reify
         expect(skipper.another_timestamp).to be(nil)
       end
@@ -35,7 +35,7 @@ RSpec.describe Skipper, type: :model, versioning: true do
     context "with preserve" do
       it "preserves its timestamp" do
         skipper = Skipper.create!(another_timestamp: t1)
-        skipper.update_attributes!(another_timestamp: t2, name: "Foobar")
+        skipper.update!(another_timestamp: t2, name: "Foobar")
         skipper = skipper.versions.last.reify(unversioned_attributes: :preserve)
         expect(skipper.another_timestamp).to eq(t2)
       end