+++ spec/paper_trail/associations/has_many_through_spec.rb
@@ -23,7 +23,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
 
     context "updated before the associated was created" do
       before do
-        @book.update_attributes!(title: "book_1")
+        @book.update!(title: "book_1")
         @book.authors.create!(name: "author_0")
       end
 
@@ -57,7 +57,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
     context "updated before it is associated with an existing one" do
       before do
         person_existing = Person.create(name: "person_existing")
-        @book.update_attributes!(title: "book_1")
+        @book.update!(title: "book_1")
         (@book.authors << person_existing)
       end
 
@@ -88,7 +88,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
       before do
         @author = @book.authors.create!(name: "author_0")
         @person_existing = Person.create(name: "person_existing")
-        @book.update_attributes!(title: "book_1")
+        @book.update!(title: "book_1")
       end
 
       context "when reified" do
@@ -101,10 +101,10 @@ RSpec.describe(::PaperTrail, versioning: true) do
 
       context "and then the associated is updated between model versions" do
         before do
-          @author.update_attributes(name: "author_1")
-          @author.update_attributes(name: "author_2")
-          @book.update_attributes(title: "book_2")
-          @author.update_attributes(name: "author_3")
+          @author.update(name: "author_1")
+          @author.update(name: "author_2")
+          @book.update(title: "book_2")
+          @author.update(name: "author_3")
         end
 
         context "when reified" do
@@ -147,7 +147,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
       context "and then the associated is destroyed between model versions" do
         before do
           @author.destroy
-          @book.update_attributes(title: "book_2")
+          @book.update(title: "book_2")
         end
 
         context "when reified" do
@@ -162,7 +162,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
       context "and then the associated is dissociated between model versions" do
         before do
           @book.authors = []
-          @book.update_attributes(title: "book_2")
+          @book.update(title: "book_2")
         end
 
         context "when reified" do
@@ -241,7 +241,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
 
     context "updated before the associated without paper_trail was created" do
       before do
-        @book.update_attributes!(title: "book_1")
+        @book.update!(title: "book_1")
         @book.editors.create!(name: "editor_0")
       end
 
@@ -259,7 +259,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
     before { @chapter = Chapter.create(name: CHAPTER_NAMES[0]) }
 
     context "before any associations are created" do
-      before { @chapter.update_attributes(name: CHAPTER_NAMES[1]) }
+      before { @chapter.update(name: CHAPTER_NAMES[1]) }
 
       it "not reify any associations" do
         chapter_v1 = @chapter.versions[1].reify(has_many: true)
@@ -271,11 +271,11 @@ RSpec.describe(::PaperTrail, versioning: true) do
 
     context "after the first has_many through relationship is created" do
       before do
-        @chapter.update_attributes(name: CHAPTER_NAMES[1])
+        @chapter.update(name: CHAPTER_NAMES[1])
         @chapter.sections.create(name: "section 1")
-        @chapter.sections.first.update_attributes(name: "section 2")
-        @chapter.update_attributes(name: CHAPTER_NAMES[2])
-        @chapter.sections.first.update_attributes(name: "section 3")
+        @chapter.sections.first.update(name: "section 2")
+        @chapter.update(name: CHAPTER_NAMES[2])
+        @chapter.sections.first.update(name: "section 3")
       end
 
       context "version 1" do
@@ -296,7 +296,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
 
       context "version 2, before the section was destroyed" do
         before do
-          @chapter.update_attributes(name: CHAPTER_NAMES[2])
+          @chapter.update(name: CHAPTER_NAMES[2])
           @chapter.sections.destroy_all
         end
 
@@ -309,7 +309,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
       context "version 3, after the section was destroyed" do
         before do
           @chapter.sections.destroy_all
-          @chapter.update_attributes(name: CHAPTER_NAMES[3])
+          @chapter.update(name: CHAPTER_NAMES[3])
         end
 
         it "have no sections" do
@@ -328,9 +328,9 @@ RSpec.describe(::PaperTrail, versioning: true) do
           it "have one paragraph" do
             initial_section_name = @section.name
             initial_paragraph_name = @paragraph.name
-            @chapter.update_attributes(name: CHAPTER_NAMES[4])
+            @chapter.update(name: CHAPTER_NAMES[4])
             expect(@chapter.versions.size).to(eq(4))
-            @paragraph.update_attributes(name: "para3")
+            @paragraph.update(name: "para3")
             chapter_v3 = @chapter.versions[3].reify(has_many: true)
             expect(chapter_v3.sections.map(&:name)).to(eq([initial_section_name]))
             paragraphs = chapter_v3.sections.first.paragraphs
@@ -341,7 +341,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
 
         context "the version before a section is destroyed" do
           it "have the section and paragraph" do
-            @chapter.update_attributes(name: CHAPTER_NAMES[3])
+            @chapter.update(name: CHAPTER_NAMES[3])
             expect(@chapter.versions.size).to(eq(4))
             @section.destroy
             expect(@chapter.versions.size).to(eq(4))
@@ -356,7 +356,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
         context "the version after a section is destroyed" do
           it "not have any sections or paragraphs" do
             @section.destroy
-            @chapter.update_attributes(name: CHAPTER_NAMES[5])
+            @chapter.update(name: CHAPTER_NAMES[5])
             expect(@chapter.versions.size).to(eq(4))
             chapter_v3 = @chapter.versions[3].reify(has_many: true)
             expect(chapter_v3.sections.size).to(eq(0))
@@ -367,7 +367,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
         context "the version before a paragraph is destroyed" do
           it "have the one paragraph" do
             initial_paragraph_name = @section.paragraphs.first.name
-            @chapter.update_attributes(name: CHAPTER_NAMES[5])
+            @chapter.update(name: CHAPTER_NAMES[5])
             @paragraph.destroy
             chapter_v3 = @chapter.versions[3].reify(has_many: true)
             paragraphs = chapter_v3.sections.first.paragraphs
@@ -379,7 +379,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
         context "the version after a paragraph is destroyed" do
           it "have no paragraphs" do
             @paragraph.destroy
-            @chapter.update_attributes(name: CHAPTER_NAMES[5])
+            @chapter.update(name: CHAPTER_NAMES[5])
             chapter_v3 = @chapter.versions[3].reify(has_many: true)
             expect(chapter_v3.paragraphs.size).to(eq(0))
             expect(chapter_v3.sections.first.paragraphs).to(eq([]))
@@ -395,7 +395,7 @@ RSpec.describe(::PaperTrail, versioning: true) do
         paragraph = Paragraph.create(name: "Paragraph One", section: section)
         quotation = Quotation.create(chapter: chapter)
         citation = Citation.create(quotation: quotation)
-        chapter.update_attributes(name: CHAPTER_NAMES[1])
+        chapter.update(name: CHAPTER_NAMES[1])
         expect(chapter.versions.count).to(eq(2))
         paragraph.destroy
         citation.destroy