+++ CHANGELOG.md
@@ -3,15 +3,28 @@
 This project follows [semver 2.0.0](http://semver.org/spec/v2.0.0.html) and the
 recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
-## Unreleased
+## 10.1.0 (Unreleased)
 
 ### Breaking Changes
 
 - None
 
+### Deprecated
+
+- [#1158](https://github.com/paper-trail-gem/paper_trail/pull/1158) - Passing
+  association name as `versions:` option or Version class name as `class_name:`
+  options directly to `has_paper_trail`. Use `has_paper_trail versions: {name:
+  :my_name, class_name: "MyVersionModel"}` instead.
+
 ### Added
 
-- None
+- [#1166](https://github.com/paper-trail-gem/paper_trail/pull/1166) -
+  New global option `has_paper_trail_defaults`, defaults for `has_paper_trail`
+- [#1158](https://github.com/paper-trail-gem/paper_trail/pull/1158) — Add the
+  ability to pass options, such as `scope` or `extend:` to the `has_many
+  :versions` association macro.
+- [#1172](https://github.com/paper-trail-gem/paper_trail/pull/1172) -
+  Support rails 6.0.0 alpha.
 
 ### Fixed
 
@@ -998,7 +1011,7 @@ in the `PaperTrail::Version` class through a `Rails::Engine` when the gem is use
     `ActionController::API` for compatibility with the [`rails-api`](https://github.com/rails-api/rails-api) gem.
   - [#312](https://github.com/paper-trail-gem/paper_trail/issues/312) - Fix RSpec `with_versioning` class level helper method.
   - `model_instance.without_versioning` now yields the `model_instance`, enabling syntax like this:
-    `model_instance.without_versioning { |obj| obj.update_attributes(:name => 'value') }`.
+    `model_instance.without_versioning { |obj| obj.update(:name => 'value') }`.
   - Deprecated `Model.paper_trail_on` and `Model.paper_trail_off` in favor of bang versions of the methods.
     Deprecation warning informs users that the non-bang versions of the methods will be removed in version `4.0`